rename resolver to dependency_manager + bug fix in cache
