+++ pyinject/overrider.py
@@ -1,17 +1,17 @@
 import typing
 
-from ._resolver import DependenciesManager, OverridesMapping, dependencies_manager
+from ._manager import DependenciesManager, OverridesMapping, dependencies_manager
 
 
 class DependencyOverrider:
-    def __init__(self, overrides: OverridesMapping, _resolver: DependenciesManager = dependencies_manager) -> None:
+    def __init__(self, overrides: OverridesMapping, _dependencies_manager: DependenciesManager = dependencies_manager) -> None:
         self.overrides = overrides
-        self._resolver = _resolver
+        self._dependencies_manager = _dependencies_manager
         self._old_overrides: OverridesMapping = {}
 
     def __enter__(self):
-        self._old_overrides = self._resolver.override_dependencies(self.overrides)
+        self._old_overrides = self._dependencies_manager.override_dependencies(self.overrides)
         return self
 
     def __exit__(self, *args: typing.Any) -> None:
-        self._resolver.restore_dependencies(self.overrides, self._old_overrides)
+        self._dependencies_manager.restore_dependencies(self.overrides, self._old_overrides)
